Finished Ass2
Yay it's done. After 3 brain burnouts and a stupid amount of
duplicated/messy code, the methods q1, q2, q3, and test1 work perfectly.
We can clean up the code if you guys like but I'm glad that at least the
answers are working. Basically there are the methods to generate a quick
path, and to generate a cheap path. Our solutions will never look for
optimization between the two. For example, for q1 where we want a trip
under 15 hours, I simply made it give the quickest path (and the most
expensive) and checked that the length of time was less then 15 hours.
For q2 that wants a cheap trip, I generated the cheapest trip (walking
everywhere) and made sure the cost was less then 2000$. Last question
asks that we have the shortest trip (distance?), which is forced by the
"cheap" and "quick" methods anyways. I decided to just use the quick
method since it'll make the shortest in terms of distance and speed.